refactor: extract debug/logging, keyboard hooks, IPC handlers, and storage migration into services/
