+++ src/main/services/storage-migration.ts
@@ -1,0 +1,44 @@
+import { dialog, BrowserWindow } from 'electron';
+import * as path from 'path';
+import fs from 'fs-extra';
+import log from 'electron-log';
+
+export function checkAndMigrateStorage(options: {
+  shouldCheck: boolean;
+  markAsked: () => void;
+  homeDir: string;
+  customDir: string;
+  win: BrowserWindow | null;
+}): void {
+  if (!options.shouldCheck) return;
+
+  const old_custom_dir = path.join(options.homeDir, '/mechvibes_custom');
+  if (!fs.existsSync(old_custom_dir)) return;
+
+  log.debug('Old custom directory exists, prompting user for migration...');
+  const response = dialog.showMessageBoxSync({
+    type: 'question',
+    buttons: ['Yes', 'Not right now', "Don't ask again"],
+    title: 'Mechvibes',
+    message: "Soundpacks have moved to a new location, do you want to migrate your old soundpacks to the new location? We'll only ask you this once.",
+    defaultId: 0,
+    cancelId: 1,
+  });
+
+  if (response === 0) {
+    log.debug('User requested migration, migrating...');
+    const oldCustomFiles = fs.readdirSync(old_custom_dir);
+    oldCustomFiles.forEach((file) => {
+      const sourcePath = path.join(old_custom_dir, file);
+      const destinationPath = path.join(options.customDir, file);
+      log.silly(`Moving ${sourcePath.replace(options.homeDir, '~')} to ${destinationPath.replace(options.homeDir, '~')}`);
+      fs.moveSync(sourcePath, destinationPath, { overwrite: true });
+    });
+    log.silly('Removing old custom directory...');
+    fs.removeSync(old_custom_dir);
+    log.debug('Migration complete.');
+    options.win?.reload();
+  } else if (response === 2) {
+    options.markAsked();
+  }
+}
